feat: 좋아요 기능에서 유효하지 않은 아이디 전달시 예외처리

+++ controllers/like.js
@@ -5,6 +5,12 @@ export const createLike = async (req, res) => {
   const { waffleCardId } = req.body;
   const userId = req.userId;
 
+  if (!isValidId(waffleCardId)) {
+    return res.status(404).json({
+      message: `와플 카드 id ${waffleCardId}은(는) 유효한 id가 아닙니다.`,
+    });
+  }
+
   const like = await likeRepository.create(userId, waffleCardId);
   res.status(201).json({ id: like.id, userId, waffleCardId });
 };
@@ -12,7 +18,12 @@ export const createLike = async (req, res) => {
 export const deleteLike = async (req, res) => {
   const { waffleCardId } = req.body;
   const userId = req.userId;
-  console.log(waffleCardId, userId);
+
+  if (!isValidId(waffleCardId)) {
+    return res.status(404).json({
+      message: `와플 카드 id ${waffleCardId}은(는) 유효한 id가 아닙니다.`,
+    });
+  }
 
   await likeRepository.remove(userId, waffleCardId);
   res.sendStatus(204);
